move v2 over to v1
