+++ Cargo.toml
@@ -14,6 +14,7 @@ include_dir = "0.7.3"
 itertools = "0.10.5"
 lazy_static = "1.4.0"
 nanoid = "0.4.0"
+ordered-float = "3.7.0"
 rand = "0.8.5"
 regex = "1.7.3"
 reqwest = { version = "0.11.16", default-features = false, features = ["json", "rustls-tls"] }
