+++ frontend/package.json
@@ -14,7 +14,7 @@
   },
   "dependencies": {
     "clsx": "^1.2.1",
-    "date-fns": "^2.29.3",
+    "date-fns": "^2.30.0",
     "fuse.js": "^6.6.2",
     "immer": "^10.0.1",
     "little-state-machine": "^4.8.0",
@@ -24,26 +24,26 @@
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.9",
     "react-icons": "^4.8.0",
-    "react-router-dom": "^6.11.0",
+    "react-router-dom": "^6.11.1",
     "tiny-invariant": "^1.3.1",
     "use-immer": "^0.9.0"
   },
   "devDependencies": {
-    "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.1",
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "@typescript-eslint/parser": "^5.59.1",
+    "@types/react": "^18.2.6",
+    "@types/react-dom": "^18.2.4",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
     "@vitejs/plugin-react-swc": "^3.3.0",
     "autoprefixer": "^10.4.14",
     "daisyui": "^2.51.6",
-    "eslint": "^8.39.0",
+    "eslint": "^8.40.0",
     "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-refresh": "^0.3.5",
+    "eslint-plugin-react-refresh": "^0.4.1",
     "postcss": "^8.4.23",
     "prettier": "^2.8.8",
     "tailwindcss": "^3.3.2",
     "typescript": "^5.0.4",
-    "vite": "^4.3.3",
+    "vite": "^4.3.5",
     "vite-plugin-eslint": "^1.8.1"
   }
 }
