+++ frontend/src/types/types.ts
@@ -39,6 +39,7 @@ export interface InitialFormState {
   source: undefined
   id: string
 }
+
 export interface LocalVideosFormState {
   source: "local-files"
   id: string
