+++ graphql/find_markers.graphql
@@ -25,6 +25,7 @@ query FindMarkersQuery(
         }
         files {
           basename
+          duration
         }
         sceneStreams {
           url
