--- src/local/mod.rs
@@ -1,2 +1,0 @@
-pub mod db;
-pub mod find;
