+++ src/main.rs
@@ -1,26 +1,21 @@
+use std::{sync::Arc, time::Duration};
+
 use axum::{
     routing::{delete, get, post},
     Router,
 };
 use color_eyre::Report;
-use compilation::generate::CompilationGenerator;
-use local::db::Database;
-use std::{sync::Arc, time::Duration};
 
-mod compilation;
-mod download_ffmpeg;
-mod error;
-mod local;
+use crate::{
+    data::database::Database, server::handlers::AppState, service::generator::CompilationGenerator,
+};
+
+mod data;
 mod server;
-mod stash;
+mod service;
 mod util;
 
 pub type Result<T> = std::result::Result<T, Report>;
-
-pub struct AppState {
-    pub generator: CompilationGenerator,
-    pub database: Database,
-}
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -39,7 +34,7 @@ async fn main() -> Result<()> {
         .with(EnvFilter::from_default_env())
         .init();
 
-    stash::config::init().await;
+    service::stash_config::init().await;
 
     let ffmpeg = CompilationGenerator::new().await?;
     let database = Database::new().await?;
@@ -49,22 +44,31 @@ async fn main() -> Result<()> {
     });
 
     let app = Router::new()
-        .route("/api/health", get(handlers::get_health))
-        .route("/api/tags", get(handlers::fetch_tags))
-        .route("/api/performers", get(handlers::fetch_performers))
-        .route("/api/scenes", get(handlers::fetch_scenes))
-        .route("/api/markers", get(handlers::fetch_markers))
-        .route("/api/clips", post(handlers::fetch_clips))
-        .route("/api/create", post(handlers::create_video))
-        .route("/api/progress", get(handlers::get_progress))
-        .route("/api/download", get(handlers::download_video))
-        .route("/api/funscript", post(handlers::get_funscript))
-        .route("/api/config", get(handlers::get_config))
-        .route("/api/config", post(handlers::set_config))
-        // .route("/api/video", post(handlers::list_videos))
-        .route("/api/video/:id", get(handlers::get_video))
-        // .route("/api/video/marker", post(handlers::persist_marker))
-        .route("/api/video/marker/:id", delete(handlers::delete_marker))
+        .route("/api/stash/health", get(handlers::stash::get_health))
+        .route("/api/stash/tags", get(handlers::stash::fetch_tags))
+        .route(
+            "/api/stash/performers",
+            get(handlers::stash::fetch_performers),
+        )
+        .route("/api/stash/scenes", get(handlers::stash::fetch_scenes))
+        .route("/api/stash/markers", get(handlers::stash::fetch_markers))
+        .route("/api/stash/config", get(handlers::stash::get_config))
+        .route("/api/stash/config", post(handlers::stash::set_config))
+        .route("/api/local/video", post(handlers::local::list_videos))
+        .route("/api/local/video/:id", get(handlers::local::get_video))
+        .route(
+            "/api/local/video/marker",
+            post(handlers::local::persist_marker),
+        )
+        .route(
+            "/api/local/video/marker/:id",
+            delete(handlers::local::delete_marker),
+        )
+        .route("/api/clips", post(handlers::common::fetch_clips))
+        .route("/api/create", post(handlers::common::create_video))
+        .route("/api/progress", get(handlers::common::get_progress))
+        .route("/api/download", get(handlers::common::download_video))
+        .route("/api/funscript", post(handlers::common::get_funscript))
         .fallback_service(static_files::service())
         .with_state(state);
