+++ src/server/dtos.rs
@@ -1,22 +1,34 @@
-use std::collections::BTreeSet;
+use std::collections::HashMap;
 
-use serde::Serialize;
+use camino::Utf8Path;
+use serde::{Deserialize, Serialize};
 
-use crate::stash::api::StashScene;
+use crate::{
+    data::{
+        database::{DbMarker, DbVideo, LocalVideoWithMarkers},
+        stash_api::{find_scenes_query::FindScenesQueryFindScenesScenes, FilterMode, StashMarker},
+    },
+    service::{
+        clip::ClipOrder,
+        generator::{find_stash_stream_url, find_stream_url, VideoResolution},
+        Clip, MarkerId, VideoId,
+    },
+};
 
 #[derive(Serialize, Debug)]
+#[serde(rename_all = "camelCase")]
 pub struct TagDto {
     pub name: String,
     pub id: String,
-    pub count: i64,
+    pub marker_count: i64,
 }
 
 #[derive(Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct PerformerDto {
-    pub name: String,
     pub id: String,
     pub scene_count: i64,
+    pub name: String,
     pub image_url: Option<String>,
     pub tags: Vec<String>,
     pub rating: Option<i64>,
@@ -25,37 +37,121 @@ pub struct PerformerDto {
 
 #[derive(Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
-pub enum VideoDto {
-    Stash {
-        id: String,
-        title: String,
-        image_url: String,
-        performers: Vec<String>,
-        marker_count: usize,
-        tags: BTreeSet<String>,
-        interactive: bool,
-        studio: Option<String>,
-        rating: Option<i64>,
-    },
-    LocalFile {
-        id: String,
-        file_name: String,
-        interactive: bool,
-    },
+pub struct MarkerDto {
+    pub id: MarkerId,
+    pub primary_tag: String,
+    pub stream_url: String,
+    pub start: f64,
+    pub end: f64,
+    pub scene_title: Option<String>,
+    pub performers: Vec<String>,
+    pub file_name: Option<String>,
+    pub scene_interactive: bool,
+    pub tags: Vec<String>,
 }
 
-impl From<StashScene> for VideoDto {
-    fn from(scene: StashScene) -> Self {
-        VideoDto::Stash {
-            id: scene.id,
-            title: scene.title,
-            image_url: "TODO".into(),
-            performers: scene.performers,
-            marker_count: scene.markers.len(),
-            tags: scene.tags.into_iter().collect(),
-            interactive: scene.interactive,
-            studio: scene.studio,
-            rating: scene.rating,
+impl From<StashMarker> for MarkerDto {
+    fn from(value: StashMarker) -> Self {
+        let stream_url = find_stash_stream_url(&value).into();
+        MarkerDto {
+            id: MarkerId::Stash(value.id),
+            stream_url,
+            primary_tag: value.primary_tag,
+            start: value.start,
+            end: value.end,
+            scene_title: value.scene_title,
+            performers: value.performers,
+            file_name: value.file_name,
+            scene_interactive: value.scene_interactive,
+            tags: value.tags,
         }
     }
+}
+
+impl From<DbMarker> for MarkerDto {
+    fn from(value: DbMarker) -> Self {
+        MarkerDto {
+            id: MarkerId::LocalFile(value.rowid.expect("must have an ID")),
+            start: value.start_time,
+            end: value.end_time,
+            file_name: Utf8Path::new(&value.file_path)
+                .file_name()
+                .map(|s| s.to_string()),
+            performers: vec![],
+            primary_tag: value.title,
+            scene_interactive: false,
+            scene_title: None,
+            stream_url: format!("TODO"),
+            tags: vec![],
+        }
+    }
+}
+
+#[derive(Serialize, Debug)]
+pub struct VideoDto {
+    // TODO
+}
+
+impl From<FindScenesQueryFindScenesScenes> for VideoDto {
+    fn from(_value: FindScenesQueryFindScenesScenes) -> Self {
+        VideoDto {}
+    }
+}
+
+impl From<DbVideo> for VideoDto {
+    fn from(_value: DbVideo) -> Self {
+        VideoDto {}
+    }
+}
+
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct SelectedMarker {
+    pub id: MarkerId,
+    pub video_id: VideoId,
+    pub selected_range: (f64, f64),
+    pub index_within_video: usize,
+}
+
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct CreateClipsBody {
+    pub clip_order: ClipOrder,
+    pub clip_duration: u32,
+    pub select_mode: FilterMode,
+    pub split_clips: bool,
+    pub markers: Vec<SelectedMarker>,
+}
+
+#[derive(Serialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct ClipsResponse {
+    pub clips: Vec<Clip>,
+    pub streams: HashMap<VideoId, String>,
+}
+
+#[derive(Serialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct ListVideoDto {
+    pub video: VideoDto,
+    pub markers: Vec<MarkerDto>,
+}
+
+impl From<LocalVideoWithMarkers> for ListVideoDto {
+    fn from(value: LocalVideoWithMarkers) -> Self {
+        ListVideoDto {
+            video: value.video.into(),
+            markers: value.markers.into_iter().map(From::from).collect(),
+        }
+    }
+}
+
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct CreateVideoBody {
+    pub file_name: String,
+    pub clips: Vec<Clip>,
+    pub markers: Vec<SelectedMarker>,
+    pub output_resolution: VideoResolution,
+    pub output_fps: u32,
 }
