+++ src/server/mod.rs
@@ -1,3 +1,4 @@
 pub mod dtos;
+pub mod error;
 pub mod handlers;
 pub mod static_files;
