+++ src/server/static_files.rs
@@ -4,7 +4,7 @@ use axum::routing::{get_service, MethodRouter};
 use include_dir::{include_dir, Dir};
 use tower_http::services::{ServeDir, ServeFile};
 
-static BUILD_DIR: Dir<'_> = include_dir!("$CARGO_MANIFEST_DIR/frontend/dist");
+static BUILD_DIR: Dir<'_> = include_dir!("frontend/dist");
 
 pub fn service() -> MethodRouter {
     let base_dir = Path::new("assets");
