+++ src/service/clip.rs
