+++ src/service/stash_config.rs
