--- src/stash/mod.rs
@@ -1,2 +1,0 @@
-pub mod api;
-pub mod config;
