--- v2/.env
@@ -1,1 +1,0 @@
-DATABASE_URL=sqlite:videos.sqlite3?mode=rwc
