--- v2/Cargo.toml
@@ -1,46 +1,0 @@
-[package]
-name = "clip-mash2"
-version = "0.5.1"
-edition = "2021"
-
-[dependencies]
-axum = { version = "0.6.15", features = ["macros"] }
-camino = { version =  "1.1.4", features = ["serde1"] }
-color-eyre = "0.6.2"
-directories = "5.0.0"
-futures = "0.3.28"
-graphql_client = "0.12.0"
-include_dir = "0.7.3"
-itertools = "0.10.5"
-lazy_static = "1.4.0"
-nanoid = "0.4.0"
-ordered-float = "3.7.0"
-rand = "0.8.5"
-regex = "1.7.3"
-reqwest = { version = "0.11.16", default-features = false, features = ["json", "rustls-tls"] }
-sanitise-file-name = "1.0.0"
-serde = { version = "1.0.160", features = ["derive"] }
-serde_json = "1.0.96"
-sqlx = { version = "0.7.0-alpha.2", features = ["runtime-tokio", "tls-rustls", "sqlite"] }
-tokio = { version = "1.27.0", features = ["full"] }
-tokio-stream = "0.1.12"
-tokio-util = { version = "0.7.7", features = ["io"] }
-tower = "0.4.13"
-tower-http = { version = "0.4.0", features = ["fs"] }
-tracing = "0.1.37"
-tracing-subscriber = { version = "0.3.16", features = ["env-filter"] }
-walkdir = "2.3.3"
-webbrowser = "0.8.9"
-
-[target.'cfg(unix)'.dependencies]
-tar = "0.4"
-xz2 = "0.1"
-
-[target.'cfg(windows)'.dependencies]
-zip = "0.6"
-
-[profile.dev.package.sqlx-macros]
-opt-level = 3
-
-[dev-dependencies]
-fake = "2.6.0"
