--- v2/src/server/mod.rs
@@ -1,4 +1,0 @@
-pub mod dtos;
-pub mod error;
-pub mod handlers;
-pub mod static_files;
